moved XML utils for handling config files to impexp-util
