+++ impexp-client-gui/src/main/java/org/citydb/gui/menu/MenuFile.java
@@ -29,7 +29,7 @@ package org.citydb.gui.menu;
 
 import com.formdev.flatlaf.extras.FlatSVGIcon;
 import org.citydb.config.Config;
-import org.citydb.config.ConfigUtil;
+import org.citydb.util.config.ConfigUtil;
 import org.citydb.config.ProjectConfig;
 import org.citydb.config.i18n.Language;
 import org.citydb.config.project.global.Logging;
