+++ impexp-config/src/main/java/org/citydb/config/ProjectConfig.java
@@ -186,7 +186,7 @@ public class ProjectConfig {
         return namespaceFilter;
     }
 
-    void setNamespaceFilter(ConfigNamespaceFilter namespaceFilter) {
+    public void setNamespaceFilter(ConfigNamespaceFilter namespaceFilter) {
         this.namespaceFilter = namespaceFilter;
     }
 }
