+++ impexp-config/src/main/java/org/citydb/config/project/query/filter/type/package-info.java
@@ -27,8 +27,8 @@
  */
 
 @javax.xml.bind.annotation.XmlAccessorType(javax.xml.bind.annotation.XmlAccessType.FIELD)
-@javax.xml.bind.annotation.XmlSchema(namespace = ConfigUtil.CITYDB_CONFIG_NAMESPACE_URI,
+@javax.xml.bind.annotation.XmlSchema(namespace = ConfigConstants.CITYDB_CONFIG_NAMESPACE_URI,
         elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
 package org.citydb.config.project.query.filter.type;
 
-import org.citydb.config.ConfigUtil;
+import org.citydb.config.util.ConfigConstants;
